+++ eggs/Naaya/eggs/Naaya/CHANGELOG.rst
@@ -1,5 +1,6 @@
 2.12.72 (unreleased)
 --------------------
+* Added SyntaxError for incorrect date format [bogdatan]
 
 2.12.71 (unreleased)
 --------------------